+++ src/main/java/com/epam/reportportal/utils/TestCaseIdUtils.java
@@ -28,6 +28,7 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Executable;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
@@ -52,7 +53,7 @@ public class TestCaseIdUtils {
 	/**
 	 * Generates a text code reference by consuming a {@link Executable}
 	 *
-	 * @param executable a executable or constructor, the value should not be null
+	 * @param executable an executable or constructor, the value should not be null
 	 * @return a text code reference
 	 */
 	@Nonnull
@@ -80,7 +81,7 @@ public class TestCaseIdUtils {
 		List<Integer> keys = new ArrayList<>();
 		for (int paramIndex = 0; paramIndex < parameterAnnotations.length; paramIndex++) {
 			for (int annotationIndex = 0; annotationIndex < parameterAnnotations[paramIndex].length;
-					annotationIndex++) {
+			     annotationIndex++) {
 				Annotation testCaseIdAnnotation = parameterAnnotations[paramIndex][annotationIndex];
 				if (testCaseIdAnnotation.annotationType() == TestCaseIdKey.class) {
 					keys.add(paramIndex);
@@ -107,7 +108,7 @@ public class TestCaseIdUtils {
 	 */
 	@Nullable
 	public static <T> TestCaseIdEntry getTestCaseId(@Nullable TestCaseId annotation, @Nullable Executable executable,
-			@Nullable List<T> parameters) {
+	                                                @Nullable List<T> parameters) {
 		return getTestCaseId(annotation, executable, null, parameters);
 	}
 
@@ -123,7 +124,7 @@ public class TestCaseIdUtils {
 	 */
 	@Nullable
 	public static <T> TestCaseIdEntry getTestCaseId(@Nullable TestCaseId annotation, @Nullable Executable executable,
-			@Nullable String codRef, @Nullable List<T> parameters) {
+	                                                @Nullable String codRef, @Nullable List<T> parameters) {
 		return getTestCaseId(annotation, executable, codRef, parameters, null);
 	}
 
@@ -140,7 +141,8 @@ public class TestCaseIdUtils {
 	 */
 	@Nullable
 	public static <T> TestCaseIdEntry getTestCaseId(@Nullable TestCaseId annotation, @Nullable Executable executable,
-			@Nullable String codRef, @Nullable List<T> parameters, @Nullable Object testInstance) {
+	                                                @Nullable String codRef, @Nullable List<T> parameters,
+	                                                @Nullable Object testInstance) {
 		if (annotation != null) {
 			if (annotation.value().isEmpty()) {
 				if (annotation.parametrized()) {
@@ -154,11 +156,12 @@ public class TestCaseIdUtils {
 			} else {
 				String idTemplate = annotation.value();
 				TemplateConfiguration templateConfig = new TemplateConfiguration(annotation.config());
-				Map<String, Object> parametersMap = ofNullable(parameters).map(params -> IntStream.range(0,
+				Map<String, Object> parametersMap = new HashMap<>();
+				ofNullable(parameters).ifPresent(params -> IntStream.range(0,
 								params.size()
 						)
 						.boxed()
-						.collect(Collectors.toMap(Object::toString, i -> (Object) params.get(i)))).orElse(null);
+						.forEach(i -> parametersMap.put(i.toString(), params.get(i))));
 				String id = TemplateProcessing.processTemplate(idTemplate,
 						testInstance,
 						executable,
@@ -181,7 +184,7 @@ public class TestCaseIdUtils {
 	}
 
 	/**
-	 * Generates Test Case ID based on a executable reference and a list of parameters
+	 * Generates Test Case ID based on an executable reference and a list of parameters
 	 *
 	 * @param <T>        parameters type
 	 * @param executable a {@link Executable} object